chore: address comments 6
🔒 Scanned for secrets using gitleaks 8.28.0
